reterminal/subsystemAlaa/controller/pyproject.toml
[project]
name = "controller"
version = "0.1.0"
description = "reTerminal device controller and FastAPI backend (A1)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "gpiozero>=2.0",
    "adafruit-circuitpython-ahtx0",
    "adafruit-extended-bus",
    "rpi-gpio>=0.7.1",
    "lgpio>=0.2.2.0",
    "python-dotenv>=1.2.2",
]
